requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf2
chromadb
faiss-cpu
langchain_google_genai
langchain_community
pytesseract